pyproject.toml
[project]
name = "baclast"
version = "0.1.1"
description = "Fast ESKAPE bacterial genome classifier using k-mer profiles"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "biopython>=1.81",
    "joblib>=1.3",
    "numpy>=1.24",
    "scikit-learn>=1.3",
]

[project.scripts]
baclast = "baclast.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
artifacts = ["baclast/model.pkl"]

[tool.hatch.build.targets.wheel]
packages = ["baclast"]

[tool.hatch.build.targets.sdist]
include = ["baclast/", "tests/", "README.md", "pyproject.toml"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ipython>=9.0",
    "matplotlib>=3.10",
    "pandas>=3.0",
]
